More informative prints, configuration is at the top of the file now, fixes few uninitialised variable errors, modify !START regex.
This commit does a few things,

First and foremost, it adds a few more informative prints, such as identifying to nickserv, and display where we are connecting to.

This also makes the configuration of the bot on the top of the file, rather than at the bottom.

There were a few uninitialised variable errors, mainly regarding empty .pwd files. We now check if the file exists, and if it has content.

Lastly, this attempts to (but doesn't do it very well, yet) modify the START game regex. However, I've come across unexpected errors that relate to $roundwordlist never being set properly, and this will have to be fixed before being able to push it to the live regolf version.

reference #28